Add CounterWithStampLock implementation and benchmarking
A new implementation of the Counter interface, CounterWithStampLock, was added. This new implementation uses a StampedLock which is a Read/Write lock that allows multiple readers, unlike standard Read/Write locks. In addition to this, CounterWithStampLock is used now in the CounterBenchmark for performance benchmarking. Updates are also made to the README.md file to reflect the changes in the benchmark results.

This allows us to compare the performance of CounterWithStampLock with the other counter implementations and choose the most optimised one for our use case.